requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.75.0
azure-identity>=1.15.0
anyio==4.10.0
attrs==25.3.0
bashlex==0.18
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
coverage==7.11.3
datasets==4.5.0
dill==0.4.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
fastapi==0.116.1
filelock==3.20.3
frozenlist==1.7.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
idna==3.10
iniconfig==2.1.0
jiter==0.11.0
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.6.4
multiprocess==0.70.18
numpy==1.26.4
openai==1.107.3
packaging==25.0
pandas==3.0.0
pexpect==4.9.0
pluggy==1.6.0
propcache==0.3.2
ptyprocess==0.7.0
pyarrow==23.0.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
swe-rex==1.4.0
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
xxhash==3.6.0
yarl==1.20.1
